[Lesson 13] cleaned up menu for bank statement app as per request
