--- bank-payments/src/main/java/lt/codeacademy/exercise/service/ShvedService.java
@@ -1,8 +1,0 @@
-package lt.codeacademy.exercise.service;
-
-public class ShvedService extends Service {
-  @Override
-  public void getFormatedData(String[][] rawData, int date, int amount, int client, int account) {
-    super.getFormatedData(rawData, 0, 2, -1, 1);
-  }
-}
